+++ src/main/java/com/koddy/server/auth/application/adapter/OAuthLoginProcessor.java
@@ -3,6 +3,7 @@ package com.koddy.server.auth.application.adapter;
 import com.koddy.server.auth.domain.model.oauth.OAuthProvider;
 import com.koddy.server.auth.domain.model.oauth.OAuthUserResponse;
 
+@FunctionalInterface
 public interface OAuthLoginProcessor {
     OAuthUserResponse login(
             final OAuthProvider provider,
