+++ src/main/java/com/koddy/server/auth/application/usecase/GetOAuthLinkUseCase.java
@@ -1,7 +1,7 @@
 package com.koddy.server.auth.application.usecase;
 
 import com.koddy.server.auth.application.usecase.query.GetOAuthLink;
-import com.koddy.server.auth.infrastructure.oauth.OAuthUriGeneratorFacade;
+import com.koddy.server.auth.infrastructure.social.OAuthUriGeneratorFacade;
 import com.koddy.server.global.annotation.UseCase;
 import lombok.RequiredArgsConstructor;
