+++ src/main/java/com/koddy/server/auth/infrastructure/social/OAuthConnector.java
@@ -1,12 +1,9 @@
-package com.koddy.server.auth.infrastructure.oauth;
+package com.koddy.server.auth.infrastructure.social;
 
-import com.koddy.server.auth.domain.model.oauth.OAuthProvider;
 import com.koddy.server.auth.domain.model.oauth.OAuthTokenResponse;
 import com.koddy.server.auth.domain.model.oauth.OAuthUserResponse;
 
 public interface OAuthConnector {
-    boolean isSupported(final OAuthProvider provider);
-
     OAuthTokenResponse fetchToken(final String code, final String redirectUri, final String state);
 
     OAuthUserResponse fetchUserInfo(final String accessToken);
