+++ src/main/java/com/koddy/server/auth/infrastructure/social/google/GoogleOAuthConnector.java
@@ -1,11 +1,10 @@
-package com.koddy.server.auth.infrastructure.oauth.zoom;
+package com.koddy.server.auth.infrastructure.social.google;
 
-import com.koddy.server.auth.domain.model.oauth.OAuthProvider;
 import com.koddy.server.auth.domain.model.oauth.OAuthTokenResponse;
 import com.koddy.server.auth.domain.model.oauth.OAuthUserResponse;
-import com.koddy.server.auth.infrastructure.oauth.OAuthConnector;
-import com.koddy.server.auth.infrastructure.oauth.zoom.response.ZoomTokenResponse;
-import com.koddy.server.auth.infrastructure.oauth.zoom.response.ZoomUserResponse;
+import com.koddy.server.auth.infrastructure.social.OAuthConnector;
+import com.koddy.server.auth.infrastructure.social.google.response.GoogleTokenResponse;
+import com.koddy.server.auth.infrastructure.social.google.response.GoogleUserResponse;
 import com.koddy.server.global.exception.GlobalException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -18,11 +17,7 @@ import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestClientException;
 import org.springframework.web.client.RestTemplate;
 
-import java.util.Base64;
-
-import static com.koddy.server.auth.domain.model.oauth.OAuthProvider.ZOOM;
 import static com.koddy.server.global.exception.GlobalExceptionCode.UNEXPECTED_SERVER_ERROR;
-import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import static org.springframework.http.HttpHeaders.CONTENT_TYPE;
 import static org.springframework.http.HttpMethod.GET;
@@ -30,14 +25,9 @@ import static org.springframework.http.HttpMethod.GET;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class ZoomOAuthConnector implements OAuthConnector {
-    private final ZoomOAuthProperties properties;
+public class GoogleOAuthConnector implements OAuthConnector {
+    private final GoogleOAuthProperties properties;
     private final RestTemplate restTemplate;
-
-    @Override
-    public boolean isSupported(final OAuthProvider provider) {
-        return provider == ZOOM;
-    }
 
     @Override
     public OAuthTokenResponse fetchToken(final String code, final String redirectUri, final String state) {
@@ -45,20 +35,13 @@ public class ZoomOAuthConnector implements OAuthConnector {
         final MultiValueMap<String, String> params = createTokenRequestParams(code, redirectUri, state);
 
         final HttpEntity<MultiValueMap<String, String>> request = new HttpEntity<>(params, headers);
-        return fetchZoomToken(request).getBody();
+        return fetchGoogleToken(request).getBody();
     }
 
     private HttpHeaders createTokenRequestHeader() {
         final HttpHeaders headers = new HttpHeaders();
         headers.set(CONTENT_TYPE, OAUTH_CONTENT_TYPE);
-        headers.set(AUTHORIZATION, String.join(" ", "Basic", getEncodedClientProperties()));
         return headers;
-    }
-
-    private String getEncodedClientProperties() {
-        final String value = String.join(":", properties.clientId(), properties.clientSecret());
-        final byte[] encode = Base64.getEncoder().encode(value.getBytes());
-        return new String(encode, UTF_8);
     }
 
     private MultiValueMap<String, String> createTokenRequestParams(
@@ -71,12 +54,14 @@ public class ZoomOAuthConnector implements OAuthConnector {
         params.add("code", code);
         params.add("redirect_uri", redirectUri);
         params.add("state", state);
+        params.add("client_id", properties.clientId());
+        params.add("client_secret", properties.clientSecret());
         return params;
     }
 
-    private ResponseEntity<ZoomTokenResponse> fetchZoomToken(final HttpEntity<MultiValueMap<String, String>> request) {
+    private ResponseEntity<GoogleTokenResponse> fetchGoogleToken(final HttpEntity<MultiValueMap<String, String>> request) {
         try {
-            return restTemplate.postForEntity(properties.tokenUrl(), request, ZoomTokenResponse.class);
+            return restTemplate.postForEntity(properties.tokenUrl(), request, GoogleTokenResponse.class);
         } catch (final RestClientException e) {
             log.error("OAuth Error... ", e);
             throw new GlobalException(UNEXPECTED_SERVER_ERROR);
@@ -87,19 +72,18 @@ public class ZoomOAuthConnector implements OAuthConnector {
     public OAuthUserResponse fetchUserInfo(final String accessToken) {
         final HttpHeaders headers = createUserInfoRequestHeader(accessToken);
         final HttpEntity<Void> request = new HttpEntity<>(headers);
-        return fetchZoomUserInfo(request).getBody();
+        return fetchGoogleUserInfo(request).getBody();
     }
 
     private HttpHeaders createUserInfoRequestHeader(final String accessToken) {
         final HttpHeaders headers = new HttpHeaders();
-        headers.set(CONTENT_TYPE, OAUTH_CONTENT_TYPE);
         headers.set(AUTHORIZATION, String.join(" ", BEARER_TOKEN_TYPE, accessToken));
         return headers;
     }
 
-    private ResponseEntity<ZoomUserResponse> fetchZoomUserInfo(final HttpEntity<Void> request) {
+    private ResponseEntity<GoogleUserResponse> fetchGoogleUserInfo(final HttpEntity<Void> request) {
         try {
-            return restTemplate.exchange(properties.userInfoUrl(), GET, request, ZoomUserResponse.class);
+            return restTemplate.exchange(properties.userInfoUrl(), GET, request, GoogleUserResponse.class);
         } catch (final RestClientException e) {
             log.error("OAuth Error... ", e);
             throw new GlobalException(UNEXPECTED_SERVER_ERROR);
