+++ src/main/java/com/koddy/server/auth/infrastructure/social/kakao/KakaoOAuthConnector.java
@@ -1,11 +1,10 @@
-package com.koddy.server.auth.infrastructure.oauth.google;
+package com.koddy.server.auth.infrastructure.social.kakao;
 
-import com.koddy.server.auth.domain.model.oauth.OAuthProvider;
 import com.koddy.server.auth.domain.model.oauth.OAuthTokenResponse;
 import com.koddy.server.auth.domain.model.oauth.OAuthUserResponse;
-import com.koddy.server.auth.infrastructure.oauth.OAuthConnector;
-import com.koddy.server.auth.infrastructure.oauth.google.response.GoogleTokenResponse;
-import com.koddy.server.auth.infrastructure.oauth.google.response.GoogleUserResponse;
+import com.koddy.server.auth.infrastructure.social.OAuthConnector;
+import com.koddy.server.auth.infrastructure.social.kakao.response.KakaoTokenResponse;
+import com.koddy.server.auth.infrastructure.social.kakao.response.KakaoUserResponse;
 import com.koddy.server.global.exception.GlobalException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -18,7 +17,6 @@ import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestClientException;
 import org.springframework.web.client.RestTemplate;
 
-import static com.koddy.server.auth.domain.model.oauth.OAuthProvider.GOOGLE;
 import static com.koddy.server.global.exception.GlobalExceptionCode.UNEXPECTED_SERVER_ERROR;
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import static org.springframework.http.HttpHeaders.CONTENT_TYPE;
@@ -27,14 +25,9 @@ import static org.springframework.http.HttpMethod.GET;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class GoogleOAuthConnector implements OAuthConnector {
-    private final GoogleOAuthProperties properties;
+public class KakaoOAuthConnector implements OAuthConnector {
+    private final KakaoOAuthProperties properties;
     private final RestTemplate restTemplate;
-
-    @Override
-    public boolean isSupported(final OAuthProvider provider) {
-        return provider == GOOGLE;
-    }
 
     @Override
     public OAuthTokenResponse fetchToken(final String code, final String redirectUri, final String state) {
@@ -42,7 +35,7 @@ public class GoogleOAuthConnector implements OAuthConnector {
         final MultiValueMap<String, String> params = createTokenRequestParams(code, redirectUri, state);
 
         final HttpEntity<MultiValueMap<String, String>> request = new HttpEntity<>(params, headers);
-        return fetchGoogleToken(request).getBody();
+        return fetchKakaoToken(request).getBody();
     }
 
     private HttpHeaders createTokenRequestHeader() {
@@ -66,9 +59,9 @@ public class GoogleOAuthConnector implements OAuthConnector {
         return params;
     }
 
-    private ResponseEntity<GoogleTokenResponse> fetchGoogleToken(final HttpEntity<MultiValueMap<String, String>> request) {
+    private ResponseEntity<KakaoTokenResponse> fetchKakaoToken(final HttpEntity<MultiValueMap<String, String>> request) {
         try {
-            return restTemplate.postForEntity(properties.tokenUrl(), request, GoogleTokenResponse.class);
+            return restTemplate.postForEntity(properties.tokenUrl(), request, KakaoTokenResponse.class);
         } catch (final RestClientException e) {
             log.error("OAuth Error... ", e);
             throw new GlobalException(UNEXPECTED_SERVER_ERROR);
@@ -79,18 +72,19 @@ public class GoogleOAuthConnector implements OAuthConnector {
     public OAuthUserResponse fetchUserInfo(final String accessToken) {
         final HttpHeaders headers = createUserInfoRequestHeader(accessToken);
         final HttpEntity<Void> request = new HttpEntity<>(headers);
-        return fetchGoogleUserInfo(request).getBody();
+        return fetchKakaoUserInfo(request).getBody();
     }
 
     private HttpHeaders createUserInfoRequestHeader(final String accessToken) {
         final HttpHeaders headers = new HttpHeaders();
+        headers.set(CONTENT_TYPE, OAUTH_CONTENT_TYPE);
         headers.set(AUTHORIZATION, String.join(" ", BEARER_TOKEN_TYPE, accessToken));
         return headers;
     }
 
-    private ResponseEntity<GoogleUserResponse> fetchGoogleUserInfo(final HttpEntity<Void> request) {
+    private ResponseEntity<KakaoUserResponse> fetchKakaoUserInfo(final HttpEntity<Void> request) {
         try {
-            return restTemplate.exchange(properties.userInfoUrl(), GET, request, GoogleUserResponse.class);
+            return restTemplate.exchange(properties.userInfoUrl(), GET, request, KakaoUserResponse.class);
         } catch (final RestClientException e) {
             log.error("OAuth Error... ", e);
             throw new GlobalException(UNEXPECTED_SERVER_ERROR);
