+++ src/main/java/com/koddy/server/auth/infrastructure/social/kakao/KakaoOAuthUriGenerator.java
@@ -1,4 +1,4 @@
-package com.koddy.server.auth.infrastructure.oauth.zoom;
+package com.koddy.server.auth.infrastructure.social.kakao;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -8,14 +8,15 @@ import java.util.UUID;
 
 @Component
 @RequiredArgsConstructor
-public class ZoomOAuthUriGenerator {
-    private final ZoomOAuthProperties properties;
+public class KakaoOAuthUriGenerator {
+    private final KakaoOAuthProperties properties;
 
     public String generate(final String redirectUri) {
         return UriComponentsBuilder
                 .fromUriString(properties.authUrl())
                 .queryParam("response_type", "code")
                 .queryParam("client_id", properties.clientId())
+                .queryParam("scope", String.join(" ", properties.scope()))
                 .queryParam("redirect_uri", redirectUri)
                 .queryParam("state", UUID.randomUUID().toString().replaceAll("-", ""))
                 .build()
