+++ src/main/java/com/koddy/server/auth/infrastructure/social/zoom/ZoomOAuthUriGenerator.java
@@ -1,4 +1,4 @@
-package com.koddy.server.auth.infrastructure.oauth.google;
+package com.koddy.server.auth.infrastructure.social.zoom;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -8,15 +8,14 @@ import java.util.UUID;
 
 @Component
 @RequiredArgsConstructor
-public class GoogleOAuthUriGenerator {
-    private final GoogleOAuthProperties properties;
+public class ZoomOAuthUriGenerator {
+    private final ZoomOAuthProperties properties;
 
     public String generate(final String redirectUri) {
         return UriComponentsBuilder
                 .fromUriString(properties.authUrl())
                 .queryParam("response_type", "code")
                 .queryParam("client_id", properties.clientId())
-                .queryParam("scope", String.join(" ", properties.scope()))
                 .queryParam("redirect_uri", redirectUri)
                 .queryParam("state", UUID.randomUUID().toString().replaceAll("-", ""))
                 .build()
