+++ src/test/java/com/koddy/server/auth/application/usecase/OAuthLoginUseCaseTest.java
@@ -6,8 +6,8 @@ import com.koddy.server.auth.domain.model.AuthMember;
 import com.koddy.server.auth.domain.model.AuthToken;
 import com.koddy.server.auth.domain.service.TokenIssuer;
 import com.koddy.server.auth.exception.OAuthUserNotFoundException;
-import com.koddy.server.auth.infrastructure.oauth.google.response.GoogleTokenResponse;
-import com.koddy.server.auth.infrastructure.oauth.google.response.GoogleUserResponse;
+import com.koddy.server.auth.infrastructure.social.google.response.GoogleTokenResponse;
+import com.koddy.server.auth.infrastructure.social.google.response.GoogleUserResponse;
 import com.koddy.server.common.UnitTest;
 import com.koddy.server.member.domain.model.Member;
 import com.koddy.server.member.domain.repository.MemberRepository;
