+++ src/test/java/com/koddy/server/common/fixture/MentorFixture.java
@@ -3,9 +3,9 @@ package com.koddy.server.common.fixture;
 import com.koddy.server.acceptance.member.MemberAcceptanceStep;
 import com.koddy.server.auth.domain.model.AuthMember;
 import com.koddy.server.auth.domain.model.AuthToken;
-import com.koddy.server.auth.infrastructure.oauth.google.response.GoogleUserResponse;
-import com.koddy.server.auth.infrastructure.oauth.kakao.response.KakaoUserResponse;
-import com.koddy.server.auth.infrastructure.oauth.zoom.response.ZoomUserResponse;
+import com.koddy.server.auth.infrastructure.social.google.response.GoogleUserResponse;
+import com.koddy.server.auth.infrastructure.social.kakao.response.KakaoUserResponse;
+import com.koddy.server.auth.infrastructure.social.zoom.response.ZoomUserResponse;
 import com.koddy.server.member.domain.model.Email;
 import com.koddy.server.member.domain.model.Language;
 import com.koddy.server.member.domain.model.mentor.Mentor;
